better handle -e 'command' if there is a shell expression.
If command contains spaces, $ or other known shell commands, it's
spawned with a "sh -c $COMMAND" instead.

this allows us to start terminology to "watch ps" or things like that.

Also changed the return code on failure to 127, similar to system(3).

NOTE: should we reset some signals? Remember that ecore traps most of them.



SVN revision: 72313