terminology can do multi instance now (in a single process). that means 2 or 3 or more windows == same terminology process. there is a checkbox to turn it on.
in theory it CAN do multipe terms within 1 windows, except there's
incomplete sizing logic code for hansling step sizing and min size
with multielp terms in the same window (imagine u laid them out in a
table grid... all we need is a table and pack n terms in and presto..
but the sizing will be "bizarre" at the moment)... if they are packed
on top of eachother like a notebook... sure - but no notebook widget
and no way to sensibly display and switch... but the code infra now
all supports it in theory. this is more about testing and making that
infra work. first will probably be a grid layout of some sort because
frankly... it's easier.

but for now... lets get this multi-instance fun on the table.



SVN revision: 81740