termio: do not handle selection if it doesn't exist.
A typical case is when you try to force xdg-open on a media (instead
of inline) by pressing "Control". This would trigger single-click box
selection on mouse-down but on mouse up the mouse didn't move yet. By
processing this as selection one also leaves the "didclick ==
EINA_TRUE" from mouse-down, leading the link to not be processed
(_activate_link() wasn't called).

PS: what an awful name is "didclick".