theme selector - use gengrid instead of list
this makes the theme selector look much nicer. it's a step in the
right direction, but not 100% there. did this more as an example for
people filling in wallpaper and color selector (colors will need to
have a scheme/theme and preview too and then a way to edit a custom
theme/scheme).