terminology: restore movie play state when the media block is reactivated.
Summary:
first phase to resolve T1075
With D786, the paused video would be in paused state even if the media block is reactivated.

But the problem is....
every backend except "gstreamer1" doesn't show any image on screen when the video is loaded in paused state. (it's working perfectly with gstreamer1)
maybe the root cause is that frame decoding is never happened yet.
This need to be investigated, in the next phase (maybe in emotion or emotion modules?)

@fix

Test Plan: terminology -> tycat -> pause video -> page down enought until the video block is deactivated -> page up -> check whether the video is in paused state or not

Reviewers: raster, billiob

CC: seoz

Maniphest Tasks: T1075

Differential Revision: https://phab.enlightenment.org/D787