Revert "Tab-switching from keyboard should exit asap. Closes T618"
This reverts commit ab3759b67b93a55e4b04dcf00497a32476457f8b.

Now the animation ends suddenly with no transition in some cases, but
smoothly in others. exit_on_sel is set when sel_exit() is called to
set it. it is ONLY called if the term to sel != current term (thus a
new term is selected immediately when the sel appears), so it can know
to begin an exit anim when the initial animation has finished. how
about instead of breaking the animatkon, just disable mouse event
handling (move, click) if exit_on_sel is set? :)