terminology: Fix issue of icon not showing under wayland
Essentially, set the proper icon_name. The icon_name should match the
name of the .desktop file so that elementary can find the fdo icon to
use. This fixes the issue of the window icon not showing up when
running under Weston. With the proper .desktop file name now set, we
no longer need to create and set an elm_win_icon_object.

Fixes T3460

@fix

Signed-off-by: Chris Michael <cp.michael@samsung.com>