add tysend cmdline and support in escapes for sending single files
this allows you to send a file via escapes to terminology which will
pop up a file save dialog and ask what to save it as and where (based
on the original name). terminology will show a progress bar too. this
is not useful locally but remotely (e.g. you ssh'd into another
machine) it's a VERY handy way of fetching a file from the remote
machine to the local machine with a display just with tysend FILE. you
can send multiple in a sequence with tysend FILE1 FILE2 FILE3 ... ...
and terminology will ask for a location and filename per file you send
(just hit cancel if you don't want to do it).

note - it needs new theme features to work. it'll fail without them.

@feature