do a thaw eval after thaw to ensure we dont miss any events
freeze stops events (in/out/down/up and so on, not del etc.) and thaw
does the oppostie. the thing is thawing to 0 freeze level never
re-evaluated everything. it was an optimization and simpler. so to
re-evaluate state and not lose thngs like in and out evets you do have
to do a thaw eval after thaw that will evaluate that state and produce
in and out events if you thaw back to 0. i don't know if this
fixes T4194 or not, but it does dot an i and cross a t.