Revert "themes: make blue more readable"
This reverts commit 5c0167c5e150df06b15008370c539c9aad19d3eb.

Porkepix on IRC pointed this out and it actually does the reverse of make
blue more readable. compare:

before:
https://framapic.org/EvV0cVyAUKUm/OZIdkLB4pOCn.png

after:
https://framapic.org/hX2gczT1jGMU/jtu3mJ3eQE2E.png

notice the bg of the bar at the bottom now become a lighter blue
making numbers hard to read. the darker blue was chosen because that
is actually the blue color used in xterm and the linux console (or
closer to it with a bit of a muted coloring to it rather than totally
pure). :( so sorry - this IMHO goes backwards and Porkepix did point
it out too.