config - font size - go back to 10 as default
so a whole back terminology moved from the shipepd nexus bitmap font
to system monospace fonts and 12 pt. it has always looked wrong as
efl's default theme is 10 and relies on a system-wide scale factor to
do this. 12 has looked totally out of place for a long time and it's
time that this not be hacked around in terminology but be solved at
the system level - maybe efl. if you use e then you will have selected
the scaling factor already and it will affect efl apps and all work.
outside of e is "not solved". to date it's been "other desktosp have
to support efl - up to them". perhaps looking at some level of "well
user has no config so let's guess what scale factor should maybe be
based on dpy and other toolkit configs we can find etc." in EFL... not
terminology.