font sizing - listen to elm config change events and re-apply font size
re-applying font size (setting it) will take int account any new scale
value that may be set. this allows elementary_config or any config
tool that is updating config to change scaling and terminology
properly adapts. elementary itself for all widgets already takes care
of its own gui elements, but since the terminal grid is a custom thing
for terminology - it is responsible for doing this itself. now it
does.

@feat