termptyesc: ignore APC escapes
Parse and discard APC control functions (defined in ECMA-48) as XTerm
does, instead of bleeding them through the output.

This is useful if a program wishes to detect support for Kitty images
through dynamic terminal querying; terminals which do not support Kitty
images are expected to ignore such APC sequences.